[TS-176] Refactor: directory name change
