+++ src/main.tsx
@@ -4,7 +4,7 @@ import ReactDOM from "react-dom/client";
 import { Global } from "@emotion/react";
 
 import App from "@/App";
-import { GlobalStyle } from "@/styles/GlobalStyle";
+import { GlobalStyle } from "@/change/GlobalStyle";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
 	<React.StrictMode>
